refactor: NamedParameterJdbcTemplate을 사용하도록 수정
